Refactor CrmCash component and update Numberr
This update refactors the CrmCash component to add logic for computing total price based on various factors. The code responsible for this calculation was extracted into a separate useEffect hook and additional input fields were added to allow the user to control these variables directly. Also, the Numberr component was updated to function component from class component, making it more consistent with modern React usage. Additionally, the proxy in package.json was changed to "http://127.0.0.1:5555" and certain views were commented out in the FormatDrukView component.